test :MineLocationValidator는 게임판의 인덱스를 벗어난 지뢰의 위치를 검증한다.
