+++ README.md
@@ -17,8 +17,8 @@
 * MineLocationGenerator
 
 - [ ] 지뢰의 위치를 생성한다.
-- [ ] 게임판의 인덱스를 벗어난 지뢰의 위치를 생성하면 IllegaArgumentException을 throw 한다.
 
 * MineLocationValidator
 
 - [x] 지뢰의 위치가 중복된 곳이 있는지 확인한다.
+- [x] 게임판의 인덱스를 벗어난 지뢰의 위치를 생성하면 IllegaArgumentException을 throw 한다.
